requirements.txt
Django==5.2.1
mysqlclient==2.2.7
environs==14.1.1
factory_boy==3.3.3
Faker==37.1.0
tdqm==4.67.1
django-debug-toolbar==5.2.0
djangorestframework==3.16.0
drf_nested_routers==0.94.2